Added abstraction for incoming atom download.
